Added first performance/correctness test. Start with: bin/gray -p perft_1

+++ trunk/src/board_base.cpp
@@ -1040,7 +1040,7 @@ void board_base::coord_to_san(move_t m, string& san)
 /*----------------------------------------------------------------------------*\
  |				    perft()				      |
 \*----------------------------------------------------------------------------*/
-int board_base::perft(int depth)
+unsigned long board_base::perft(int depth)
 {
 
 // From the current position, grow the move tree to the given depth and count
@@ -1049,7 +1049,7 @@ int board_base::perft(int depth)
 
 	list<move_t> l;
 	list<move_t>::iterator it;
-	int nodes = 0;
+	unsigned long nodes = 0;
 
 	// Base case.
 	if (depth == 0)
